Add files via upload
The Association for Computing Machinery Women in short for ACMW supports, celebrates, and advocates internationally for the full engagement of women in all aspects of the computing field, providing a wide range of programs and services to ACM members and working in the larger community to advance the contributions of technical women. ACM-W is an active organization with over 36,000 members. This project is about the Events and services of the ACMW body of our college. The details of various events that are held throughout many academic years in our college are displayed in this project. The technical stack used for completing this project are mainly HTML & CSS, JavaScript.